pyproject.toml
[project]
name = "tpa-form-validator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elsai-docloom==0.1.4",
    "gradio>=6.10.0",
    "huggingface-hub[cli]>=1.8.0",
    "python-dotenv>=1.2.2",
]

[[tool.uv.index]]
url = "https://elsai-core-package.optisolbusiness.com/root/elsai-docloom/"
name = "elsai-docloom"
